+++ app/src/main/java/name/eraxillan/airinganimeschedule/db/RemoteKeys.kt
@@ -21,7 +21,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "remote_keys")
 data class RemoteKeys(
-    @PrimaryKey val anilistId: Int,
+    @PrimaryKey val anilistId: Long,
     val prevKey: Int?,
     val nextKey: Int?
 )
