+++ app/src/main/java/name/eraxillan/airinganimeschedule/model/AiringAnime.kt
@@ -103,7 +103,7 @@ enum class AiringAnimeFormat {
 @Parcelize
 data class AiringAnime constructor(
     @PrimaryKey
-    var anilistId: Int = -1,          // Anilist identifier
+    var anilistId: Long = -1,         // Anilist identifier
     var url: URL = INVALID_URL,       // Anilist/MyAnimeList/Wakanim/Crunchyroll/etc. URL
     var season: Int = -1,             // Season number
     var format: AAF = AAF.TV,         // The format the media was released in
@@ -124,5 +124,5 @@ data class AiringAnime constructor(
 @Parcelize
 data class FavoriteAnime constructor(
     @PrimaryKey
-    val anilistId: Int = -1,          // Anilist identifier
+    val anilistId: Long = -1,          // Anilist identifier
 ) : Parcelable
